added new endpoints

+++ backend/src/main/java/com/pwr/edu/backend/service/PizzaService.java
@@ -2,7 +2,9 @@ package com.pwr.edu.backend.service;
 
 import com.pwr.edu.backend.domain.dto.PizzaDto;
 import com.pwr.edu.backend.domain.pizza.Pizza;
+import com.pwr.edu.backend.domain.security.ConfirmationToken;
 import com.pwr.edu.backend.repository.PizzaRepository;
+import com.pwr.edu.backend.repository.security.ConfirmationTokenRepository;
 import com.pwr.edu.backend.service.calculation.PriceCalculator;
 import com.pwr.edu.backend.service.calculation.SwitchCalculationStrategy;
 import lombok.AllArgsConstructor;
@@ -11,31 +13,25 @@ import org.springframework.transaction.annotation.Transactional;
 import com.pwr.edu.backend.exceptions.NotFoundException;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
 public class PizzaService {
     private final PizzaRepository pizzaRepository;
+    private final ConfirmationTokenRepository confirmationTokenRepository;
 
     @Transactional
-    public Pizza createPizza(Pizza pizza) {
+    public Pizza createPizza(Pizza pizza, String jwt) {
         PriceCalculator priceCalculator = new PriceCalculator(new SwitchCalculationStrategy());
 
-        Pizza createdPizza = new Pizza(pizza.getName(), pizza.getDough(), pizza.getSauce(),
-                pizza.getSize(), pizza.getIngredientsList());
-
-        pizzaRepository.save(createdPizza);
-
-        Pizza neededPizzaForPriceUpdate = pizzaRepository.findPizzaByName(createdPizza.getName())
-                .orElseThrow(NotFoundException::new);
-
-        int price = priceCalculator.calculatePrice(createdPizza);
-
-        pizzaRepository.updatePizzaPrice(price, neededPizzaForPriceUpdate.getId());
-
-
-        return neededPizzaForPriceUpdate;
+        int price = priceCalculator.calculatePrice(pizza);
+        ConfirmationToken confirmationToken = confirmationTokenRepository.findByToken(jwt).orElseThrow(NotFoundException::new);
+        pizza.setUser(confirmationToken.getAppUser());
+        pizzaRepository.save(pizza);
+        pizzaRepository.updatePizzaPrice(price, pizza.getId());
+        return pizza;
 
     }
 
@@ -52,8 +48,17 @@ public class PizzaService {
         return pizzaToPizzaDto(wantedPizza);
     }
 
-    private PizzaDto pizzaToPizzaDto(Pizza pizza){
+    public List<PizzaDto> findUsersAllPizzas(String jwt) {
+        ConfirmationToken confirmationToken = confirmationTokenRepository.findByToken(jwt).orElseThrow(NotFoundException::new);
+
+        return pizzaRepository.findPizzasByUserEmail(confirmationToken.getAppUser().getEmail())
+                .stream().map(this::pizzaToPizzaDto).collect(Collectors.toList());
+
+    }
+
+    private PizzaDto pizzaToPizzaDto(Pizza pizza) {
         return PizzaDto.builder()
+                .appUser(pizza.getUser())
                 .dough(pizza.getDough())
                 .size(pizza.getSize())
                 .sauce(pizza.getSauce())
@@ -63,7 +68,14 @@ public class PizzaService {
                 .id(pizza.getId())
                 .build();
     }
-    public void deletePizzaById(Long id) {
-        pizzaRepository.deleteById(id);
+
+    public void deletePizzaById(Long id, String jwt) {
+        Pizza pizza = pizzaRepository.findById(id).orElseThrow(NotFoundException::new);
+        ConfirmationToken confirmationToken = confirmationTokenRepository.findByToken(jwt).orElseThrow(NotFoundException::new);
+        if (Objects.equals(pizza.getUser().getEmail(), confirmationToken.getAppUser().getEmail())) {
+            pizzaRepository.deleteById(id);
+        } else {
+            throw new IllegalStateException("You cannot delete not your own pizza");
+        }
     }
 }
